gp-backend/requirements.txt
# Flask and extensions
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1

# Database
SQLAlchemy>=2.0.25

# Environment variables
python-dotenv==1.0.0

# Utilities
werkzeug==3.0.1

# AI & Agents
langchain
langchain-community
langchain-openai
openai
pydantic>=2.0
